+++ src/repository/postgres/startup-seeders/apy-seeder.js
@@ -1,17 +1,11 @@
-const { C } = require('../../../constants/runtime-constants');
-const SiloApyService = require('../../../service/silo-apy');
+const YieldService = require('../../../service/yield-service');
 const Concurrent = require('../../../utils/async/concurrent');
 const Log = require('../../../utils/logging');
-const BeanstalkSubgraphRepository = require('../../subgraph/beanstalk-subgraph');
-const YieldRepository = require('../queries/yield-repository');
 
 class ApySeeder {
   static async run() {
-    const currentSeason = (await BeanstalkSubgraphRepository.getLatestSeason()).season;
-
     // Find all missing seasons
-    let missingSeasons = await YieldRepository.findMissingSeasons(currentSeason);
-    missingSeasons = missingSeasons.filter((s) => s >= C().MIN_EMA_SEASON);
+    let missingSeasons = await YieldService.findMissingSeasons();
 
     // Calculate and save all vapys for each season (this will take a long time)
     // Currently Pre-exploit seasons are expected to fail
@@ -19,8 +13,8 @@ class ApySeeder {
     for (const season of missingSeasons) {
       try {
         // TODO: This can be increased once the decentralized subgraphs are deployed + the rate limit is increased
-        await Concurrent.run(TAG, 5, async () => {
-          await SiloApyService.saveSeasonalApys({ season });
+        await Concurrent.run(TAG, 2, async () => {
+          await YieldService.saveSeasonalApys({ season });
         });
         Log.info(`Saved apy for season ${season}`);
       } catch (e) {
