+++ src/service/silo-apy.js
@@ -2,9 +2,6 @@
  * @typedef {import('../../types/types').GetApyRequest} GetApyRequest
  * @typedef {import('../../types/types').CalcApysResult} CalcApysResult
  * @typedef {import('../../types/types').WindowEMAResult} WindowEMAResult
- *
- * @typedef {import('../../types/types').GetApyHistoryRequest} GetApyHistoryRequest
- * @typedef {import('../../types/types').GetApyHistoryResult} GetApyHistoryResult
  */
 
 const BeanstalkSubgraphRepository = require('../repository/subgraph/beanstalk-subgraph');
@@ -17,8 +14,6 @@ const { ApyInitType } = require('../repository/postgres/models/types/types');
 const YieldModelAssembler = require('../repository/postgres/models/assemblers/yield-assembler');
 const RestParsingUtil = require('../utils/rest-parsing');
 const { C } = require('../constants/runtime-constants');
-const TokenRepository = require('../repository/postgres/queries/token-repository');
-const AsyncContext = require('../utils/async/context');
 
 class SiloApyService {
   static DEFAULT_WINDOWS = [24, 168, 720];
@@ -53,23 +48,6 @@ class SiloApyService {
     let { season, emaWindows, tokens, options } = request;
     tokens = tokens.map((t) => t.toLowerCase());
     return await this.calcApy(season, emaWindows, tokens, options);
-  }
-
-  /**
-   * Gets the requested historical vAPYs from the database, if they have been calculated already
-   * @param {GetApyHistoryRequest} request
-   * @returns {Promise<GetApyHistoryResult>}
-   */
-  static async getHistoricalApy(request) {
-    const yieldModels = await YieldRepository.findHistoricalYields(request);
-    return yieldModels.reduce((acc, next) => {
-      acc[next.season] = {
-        bean: next.beanYield,
-        stalk: next.stalkYield,
-        ownership: next.ownershipYield
-      };
-      return acc;
-    }, {});
   }
 
   /**
@@ -288,42 +266,6 @@ class SiloApyService {
       });
     }
     return windowResults;
-  }
-
-  // Computes and saves the deafult seasonal apy entries.
-  // Defaults to the latest season if none is provided
-  static async saveSeasonalApys({ season, tokenModels } = {}) {
-    if (!season) {
-      season = (await BeanstalkSubgraphRepository.getLatestSeason()).season;
-    }
-
-    if (!tokenModels) {
-      // Determine list of tokens active at this point from subgraph
-      const availableTokens = await BeanstalkSubgraphRepository.getPreviouslyWhitelistedTokens({ season }, C(season));
-      tokenModels = await TokenRepository.findByAddresses(availableTokens.all);
-    }
-    const tokenAddrs = tokenModels.map((t) => t.address.toLowerCase());
-
-    // Compute with both init types. Skips `getApy` method - avoids db check and rest validations
-    const [latestAvgApy, latestNewApy] = await Promise.all([
-      SiloApyService.calcApy(season, SiloApyService.DEFAULT_WINDOWS, tokenAddrs, {
-        initType: ApyInitType.AVERAGE
-      }),
-      SiloApyService.calcApy(season, SiloApyService.DEFAULT_WINDOWS, tokenAddrs, {
-        initType: ApyInitType.NEW
-      })
-    ]);
-
-    // Prepare rows
-    const yieldRows = [
-      ...YieldModelAssembler.toModels(latestAvgApy, ApyInitType.AVERAGE, tokenModels),
-      ...YieldModelAssembler.toModels(latestNewApy, ApyInitType.NEW, tokenModels)
-    ];
-
-    // Save new yields
-    await AsyncContext.sequelizeTransaction(async () => {
-      return await YieldRepository.addYields(yieldRows);
-    });
   }
 }
